requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydub>=0.25.1
python-multipart>=0.0.9
python-dotenv>=1.0.1
pytest>=8.2.0
ag-ui-adk
google-adk

google-cloud-storage>=2.18.0
